# Tude-Web-App-master/backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Étude+ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
        role:
          type: string
          enum: [student, professor, admin]
        fullName:
          type: string
        profilePhoto:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        createdAt:
          type: string
        studentProfile:
          $ref: "#/components/schemas/StudentProfile"
          nullable: true
        professorProfile:
          $ref: "#/components/schemas/Professor"
          nullable: true
      required: [id, email, role, fullName, createdAt]
    StudentProfile:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        gradeLevel:
          type: string
          nullable: true
        schoolName:
          type: string
          nullable: true
        preferredSubjects:
          type: array
          items:
            type: string
        parentContact:
          type: string
          nullable: true
      required: [id, userId, preferredSubjects]
    Professor:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        fullName:
          type: string
        profilePhoto:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        subjects:
          type: array
          items:
            type: string
        gradeLevels:
          type: array
          items:
            type: string
        yearsOfExperience:
          type: number
          nullable: true
        bio:
          type: string
          nullable: true
        qualifications:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        rating:
          type: number
          nullable: true
        totalReviews:
          type: number
        totalStudents:
          type: number
        isVerified:
          type: boolean
        createdAt:
          type: string
      required: [id, userId, fullName, subjects, gradeLevels, status, totalReviews, totalStudents, isVerified, createdAt]
...
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        city:
          type: string
        profilePhoto:
          type: string
        gradeLevel:
          type: string
        schoolName:
          type: string
        preferredSubjects:
          type: array
          items:
            type: string
        subjects:
          type: array
          items:
            type: string
        gradeLevels:
          type: array
          items:
            type: string
        yearsOfExperience:
          type: number
        bio:
          type: string
        qualifications:
          type: string
...
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required: [users, total, page, limit]
...